fix: resolve Compose Material3 API compatibility issues
- Fix Int to Long type mismatch in CommandExecutor.kt (timeoutSeconds)
- Fix OutlinedTextField suffixIcon -> trailingIcon in CommandsScreen & ServicesScreen
- Fix Switch onValueChange -> onCheckedChange
- Fix Button color/contentColor -> colors with ButtonDefaults.buttonColors
- Fix MaterialTheme.colorScheme.success -> primary (not available in M3 1.1.2)
- Fix Row horizontalArrangement syntax (modifier chaining -> direct params)
- Fix horizontalArrangement = Arrangement.End -> horizontalArrangement(Arrangement.End)
- Add missing imports (background, clickable, Color, RoundedCornerShape)
- Fix mutableStateOf type inference with explicit generic types in SettingsScreen
- Fix Button color/contentColor parameters to use colors parameter

+++ app/src/main/java/com/example/linuxtermuxpanel/execution/CommandExecutor.kt
@@ -12,7 +12,7 @@ class CommandExecutor(
     init {
         val termuxPackageName = settings.termuxPackageName.ifEmpty { "com.termux" }
         val timeoutSeconds = if (settings.timeoutSeconds > 0) settings.timeoutSeconds else 30
-        termuxExecutor = FileBasedTermuxExecutor(context, termuxPackageName, timeoutSeconds)
+        termuxExecutor = FileBasedTermuxExecutor(context, termuxPackageName, timeoutSeconds.toLong())
     }
 
     override suspend fun execute(command: String): ExecutionResult =
